Refactor: create infinite scroll custom hook
