+++ src/components/Goals/GoalList/index.tsx
@@ -1,41 +1,16 @@
 'use client';
 
-import { useEffect, useRef } from 'react';
-
 import { ProgressLine } from '@/components/Dashboard/GoalList/GoalItem/ProgressLine';
 import { TodoList } from '@/components/Dashboard/GoalList/GoalItem/TodoList';
 
 import { useGoalsDetailQuery } from '@/hooks/apis/Goals/useGoalsDetailQuery';
 
+import { useInfiniteScroll } from '@/hooks/useInfiniteScroll';
 import { GoalHeader } from './GoalHeader';
 
 export const GoalList = () => {
-  const observerRef = useRef<HTMLDivElement | null>(null);
-
   const { goals, fetchNextPage, isLoading } = useGoalsDetailQuery();
-
-  useEffect(() => {
-    if (!observerRef.current) return;
-
-    const observerCallback: IntersectionObserverCallback = (entries) => {
-      const lastEntry = entries[0];
-      if (lastEntry.isIntersecting && !isLoading) {
-        fetchNextPage();
-      }
-    };
-
-    const observer = new IntersectionObserver(observerCallback, {
-      threshold: 1.0,
-    });
-
-    if (observerRef.current) {
-      observer.observe(observerRef.current);
-    }
-
-    observer.observe(observerRef.current);
-
-    return () => observer.disconnect();
-  }, [fetchNextPage, isLoading]);
+  const { observerRef } = useInfiniteScroll({ fetchNextPage, isLoading });
 
   return (
     <div className="flex flex-col gap-16">
